pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.3",
    "djangorestframework>=3.15.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
